Use a "common" webpack configuration Use "webpack-merge" and add source maps
