+++ package.json
@@ -21,11 +21,11 @@
   "license": "MIT",
   "main": "webpack.config.js",
   "scripts": {
-    "build:prod": "webpack --env=prod -p --config webpack.config.js --progress",
-    "build:dev": "webpack --env=dev --config webpack.config.js --progress",
+    "build:prod": "webpack --env=prod -p --progress",
+    "build:dev": "webpack --env=dev --progress",
     "lint": "node ./node_modules/eslint/bin/eslint 'src/**/*.js'",
     "lint-fix": "node ./node_modules/eslint/bin/eslint 'src/**/*.js' --fix",
-    "serve": "webpack-dev-server --env=dev --config webpack.config.js --open"
+    "serve": "webpack-dev-server --env=dev --open"
   },
   "dependencies": {},
   "devDependencies": {
@@ -48,6 +48,7 @@
     "style-loader": "^0.18.2",
     "uglify-js": "^3.0.27",
     "webpack": "^3.5.4",
-    "webpack-dev-server": "^2.7.1"
+    "webpack-dev-server": "^2.7.1",
+    "webpack-merge": "^4.1.0"
   }
 }
