+++ webpack.dev.js
@@ -1,79 +1,16 @@
-const babelPresets = ['es2015', 'es2016', 'es2017'];
+const commonConfig = require('./webpack.common.js');
+const merge = require('webpack-merge');
 const path = require('path');
 const webpack = require('webpack');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-module.exports = {
+module.exports = merge(commonConfig, {
   devServer: {
     contentBase: path.join(__dirname, 'dist'),
-    compress: true,
     port: 9000
   },
-  entry: {
-    app: './src/app.js',
-    print: './src/print.js'
-  },
-  output: {
-    filename: '[name].[chunkhash].js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        include: /src/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: babelPresets,
-            cacheDirectory: 'cache'
-          }
-        }
-      },
-      {
-        test: /\.less$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: [{
-            loader: 'css-loader'
-          }, {
-            loader: 'less-loader'
-          }]
-        })
-      },
-      {
-        test: /\.css$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader']
-        })
-      },
-      {
-        test: /\.(gif|jpg|png|svg)$/,
-        use: [
-          'file-loader'
-        ]
-      },
-      {
-        test: /\.(eot|otf|ttf|woff|woff2)$/,
-        use: [
-          'file-loader'
-        ]
-      }
-    ]
-  },
+  devtool: 'inline-source-map',
   plugins: [
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'common'
-    }),
-    new webpack.LoaderOptionsPlugin({
-      minimize: false,
-      debug: true
-    }),
     new webpack.optimize.UglifyJsPlugin({
       beautify: true,
       mangle: {
@@ -85,24 +22,9 @@ module.exports = {
       },
       comments: true
     }),
-    new CleanWebpackPlugin(['dist']),
-    new ExtractTextPlugin('[chunkhash].css'),
-    new FaviconsWebpackPlugin({
-      logo: './assets/images/pingoo.png',
-      persistentCache: true,
-      background: '#3737c8',
-      icons: {
-        android: true,
-        appleIcon: true,
-        appleStartup: true,
-        coast: false,
-        favicons: true,
-        firefox: true,
-        opengraph: false,
-        twitter: false,
-        yandex: false,
-        windows: true
-      }
+    new webpack.LoaderOptionsPlugin({
+      minimize: false,
+      debug: true
     }),
     new HtmlWebpackPlugin({
       inject: false,
@@ -118,4 +40,4 @@ module.exports = {
       title: 'My ES6 application skeleton'
     })
   ]
-};
+});
