+++ webpack.prod.js
@@ -1,70 +1,10 @@
-const babelPresets = ['es2015', 'es2016', 'es2017'];
-const path = require('path');
+const commonConfig = require('./webpack.common.js');
+const merge = require('webpack-merge');
 const webpack = require('webpack');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-module.exports = {
-  entry: {
-    app: './src/app.js',
-    print: './src/print.js'
-  },
-  output: {
-    filename: '[name].[chunkhash].js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        include: /src/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: babelPresets,
-            cacheDirectory: 'cache'
-          }
-        }
-      },
-      {
-        test: /\.less$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: [{
-            loader: 'css-loader'
-          }, {
-            loader: 'less-loader'
-          }]
-        })
-      },
-      {
-        test: /\.css$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader']
-        })
-      },
-      {
-        test: /\.(gif|jpg|png|svg)$/,
-        use: [
-          'file-loader'
-        ]
-      },
-      {
-        test: /\.(eot|otf|ttf|woff|woff2)$/,
-        use: [
-          'file-loader'
-        ]
-      }
-    ]
-  },
+module.exports = merge(commonConfig, {
   plugins: [
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'common'
-    }),
     new webpack.optimize.UglifyJsPlugin({
       beautify: false,
       mangle: {
@@ -76,24 +16,9 @@ module.exports = {
       },
       comments: false
     }),
-    new CleanWebpackPlugin(['dist']),
-    new ExtractTextPlugin('[chunkhash].css'),
-    new FaviconsWebpackPlugin({
-      logo: './assets/images/pingoo.png',
-      persistentCache: true,
-      background: '#3737c8',
-      icons: {
-        android: true,
-        appleIcon: true,
-        appleStartup: true,
-        coast: false,
-        favicons: true,
-        firefox: true,
-        opengraph: false,
-        twitter: false,
-        yandex: false,
-        windows: true
-      }
+    new webpack.LoaderOptionsPlugin({
+      minimize: true,
+      debug: false
     }),
     new HtmlWebpackPlugin({
       inject: false,
@@ -113,4 +38,4 @@ module.exports = {
       title: 'My ES6 application skeleton'
     })
   ]
-};
+});
